requirements.txt
tensorflow
tensorflow-addons
numpy
pyyaml